完成基本UI，QQ音乐插件除部分ape格式外已经全部支持

+++ BuildInMusicSources/QQMusic/QQMusicSearchResult.cs
@@ -179,7 +179,7 @@ namespace QQMusic
             public Mv1 mv { get; set; }
             public string name { get; set; }
             public int newStatus { get; set; }
-            public int nt { get; set; }
+            public long nt { get; set; }
             public Pay1 pay { get; set; }
             public int pure { get; set; }
             public Singer[] singer { get; set; }
